requirements.txt
asgiref==3.9.1
crispy-bootstrap4==2025.6
Django==5.2.6
django-cors-headers==4.7.0
django-crispy-forms==2.4
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
